LICENSE: change to Apache 2.0
GPLv3 is just an unfortunate error here. tzhash is a library and the policy we
have is that libraries are Apache 2.0, GPLv3 is used for programs. SDK
directly imports this library, so technically before this change SDK can be
considered as GPLv3 as well which is not what we want.

The code in this repository is completely written by Morphbits employees (zero
external contributions) that signed appropriate IP transfer documents,
therefore Morphbits can change the license in this case and fix the issue.

Signed-off-by: Roman Khimov <roman@nspcc.ru>